Import Debian version 4.20201102+nmu1
dh-python (4.20201102+nmu1) unstable; urgency=medium
.
  * Non-maintainer upload
  * Don't fail tests just because the default python gets hardcoded
    (Closes: #975536)
.
dh-python (4.20201102) unstable; urgency=medium
.
  [ Antonio Terceiro ]
  * pybuild's Flit plugin: don't crash on ill-formed pyproject.toml
.
  [ Piotr Ożarowski ]
  * Remove now deprecated allow_hosts=None from autogenerated .pydistutils.cfg
    file (closes: 97052)
  * Try to build/test other interpreters/versions even if one of them fails to
    make build logs a bit more verbose (closes: 943334)
  * Remove .pytest_cache and .coverage in pybuild's clean step (closes: 967898)
    Custom plugins now invoke Base.clean() as well (i.e. PyTest files,
    build dir and all .pyc files are removed like in other plugins)
.
dh-python (4.20200925) unstable; urgency=medium
.
  * Add Breaks/Replaces python2 to ease upgrades with python2 (closes: 968046)
  * Disable Python 2.X tests by default (closes: 968193)
.
dh-python (4.20200804) unstable; urgency=medium
.
  [ Scott Kitterman ]
  * Add new pybuild plugin to enable flit built packages to build using
    pyproject.toml instead of distutils/setuptools setup.py
  * Update dh-python for removal of python3.7 from supported python3 versions
    (closes: #956190)
    - Drop python3.7 from dhpython/_versions.py supported versions
    - Disable obsolete doctest in dhpython/interpreter.py
.
  [ Piotr Ożarowski ]
  * pybuild.pm: add support for python2.7 and python2.7-dbg in Build-Depends
    (closes: 949372), note that python2 and python2-dbg are not supported
  * ship Python 2's debhelper files (postinst-pycompile, prerm-pyclean and
    python2.pm) now that python package is no longer available (closes: 966832)
  * Standards-Version is 4.5.0 now (no changes needed)
.
dh-python (4.20200315) unstable; urgency=medium
.
  [ Piotr Ożarowski ]
  * Adjust for Python 3.8
    - fix setting _PYTHON_SYSCONFIGDATA_NAME (closes: 948320)
    - update tests (closes: 952122)
  * Add dh-python to Depends in autopkgtest (closes: 947789)
  * pybuild: add support for nose2. Enable by adding python3-nose2 to
    Build-Depends or --test-nose2 option or PYBUILD_TEST_NOSE2=1 env. var.
    (closes: 950912)
.
  [ Simon Ruggier ]
  * Fix a problem in handling ranges with no minimum
.
  [ Matthias Klose ]
  * Fix setting _PYTHON_HOST_PLATFORM (closes: 948319)
.
dh-python (4.20191017) unstable; urgency=medium
.
  * dh_python2: generate python2 dependencies instead of python ones
    (to make "python" package removal a bit easier)
  * generate pylint dependencies instead of pylint3 (Python 3.X files moved to
    new binary package)
.
dh-python (4.20190722) unstable; urgency=medium
.
  * Build-depend on python3-docutils to build manpages
    (reintroduces circular dependencies problem, but makes Python 2.X removal
    a bit easier)
  * Use debhelper compat level 12
  * Standards-Version is 4.4.0 now (no changes needed)

+++ dhpython/fs.py
@@ -65,15 +65,16 @@ def fix_locations(package, interpreter, versions, options):
                     pass
 
         # move files from /usr/include/pythonX.Y/ to …/pythonX.Ym/
-        srcdir = "debian/%s%s" % (package, interpreter.symlinked_include_dir)
-        if srcdir and isdir(srcdir):
-            dstdir = "debian/%s%s" % (package, interpreter.include_dir)
-            log.debug('moving files from %s to %s', srcdir, dstdir)
-            share_files(srcdir, dstdir, interpreter, options)
-            try:
-                os.removedirs(srcdir)
-            except OSError:
-                pass
+        if interpreter.symlinked_include_dir:
+            srcdir = "debian/%s%s" % (package, interpreter.symlinked_include_dir)
+            if srcdir and isdir(srcdir):
+                dstdir = "debian/%s%s" % (package, interpreter.include_dir)
+                log.debug('moving files from %s to %s', srcdir, dstdir)
+                share_files(srcdir, dstdir, interpreter, options)
+                try:
+                    os.removedirs(srcdir)
+                except OSError:
+                    pass
 
 
 def share_files(srcdir, dstdir, interpreter, options):
@@ -182,8 +183,7 @@ class Scan:
                 self.current_private_dir = self.check_private_dir(root)
                 if not self.current_private_dir:
                     # i.e. not a public dir and not a private dir
-                    is_bin_dir = self.is_bin_dir(root)
-                    if is_bin_dir:
+                    if self.is_bin_dir(root):
                         self.handle_bin_dir(root, file_names)
                     else:  # not a public, private or bin directory
                         # continue with a subdirectory
